requirements.txt
annotated-types==0.6.0
anthropic==0.25.8
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
cryptography==42.0.5
cssselect==1.2.0
distlib==0.3.8
distro==1.9.0
docopt==0.6.2
eutils==0.6.0
filelock==3.13.4
fsspec==2024.3.1
h11==0.14.0
habanero==1.2.6
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
iniconfig==2.0.0
Levenshtein==0.25.1
lxml==5.2.1
lxml_html_clean==0.1.1
metapub==0.5.5
openai==1.27.0
packaging==24.0
pdfminer.six==20231228
platformdirs==4.2.1
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pyproject-api==1.6.1
pytest==8.2.0
python-Levenshtein==0.25.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
tokenizers==0.19.1
tox==4.15.0
tqdm==4.66.2
typing_extensions==4.11.0
Unidecode==1.3.8
urllib3==2.2.1
virtualenv==20.26.0
